pyproject.toml
[project]
name = "binancebot"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests",
    "pandas",
    "matplotlib",
    "loguru",
    "binance-connector",
    "python-binance",
    "black",
    "mypy",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
